CreateIObjects now works. More simplification and tidy up.
