+++ Speckle_Adapter/Speckle_Adapter.csproj
@@ -150,8 +150,10 @@
     <Compile Include="AdapterActions\Pull.cs" />
     <Compile Include="AdapterActions\Push.cs" />
     <Compile Include="AdapterActions\_PushMethods\PushByType.cs" />
-    <Compile Include="AdapterActions\_PushMethods\ConfigureHistory.cs" />
-    <Compile Include="CRUD\Create.cs" />
+    <Compile Include="AdapterActions\_PushMethods\SetupHistory.cs" />
+    <Compile Include="CRUD\Create\CreateIObjects.cs" />
+    <Compile Include="CRUD\Create\CreateIBHoMObjects.cs" />
+    <Compile Include="CRUD\Create\CreateObjects.cs" />
     <Compile Include="CRUD\Delete.cs" />
     <Compile Include="CRUD\Read.cs" />
     <Compile Include="SpeckleAdapter.cs" />
